Major updates now the user can update the ambitions and get like-realtime updates in their ambitions view page, the users can now update ambitions in an even better UX.
Users can now delete their accounts, choosing that option has proper prompts and features are implemented in a structured manner, deleting the accounts, deletes all the ambitions that are associated by that user as well.

All ambitions page is updated with resolved bugs.

Now, creating new ambitions is better than before and ambitionStatus logic is updated to be determined by the states of ambitionTasks.

Added some animations in the tailwind config that are used in the app when the user is in the update mode in the app to update the ambitions...